not working (has memory leaks)

+++ main.c
@@ -31,7 +31,7 @@ static char *read_file(const char *path) {
     }
 
     fseek(file, 0L, SEEK_END);
-    size_t file_size = ftell(file);
+    size_t file_size = (size_t)ftell(file);
     rewind(file);
 
     char *buffer = (char *)malloc(file_size + 1);
